# OneMind/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OneMind API",
    "version": "1.0.0",
    "description": "OneMind Bun Backend API - Auto-generated from endpoint schemas",
    "contact": {
      "name": "OneMind"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Development"
    }
  ],
...
  "components": {
    "schemas": {
...
      "DeletedResponse": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "deleted"
        ]
      },
...
      "ChatListResponse": {
        "type": "object",
        "properties": {
          "chats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMetadata"
            }
          }
        },
        "required": [
          "chats"
        ]
      },
      "ChatMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "01935c4a-1b2d-7e3f-8c9d-0e1f2a3b4c5d",
            "description": "UUIDv7 identifier"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid",
            "example": "01935c4a-1b2d-7e3f-8c9d-0e1f2a3b4c5d",
            "description": "UUIDv7 identifier"
          },
          "title": {
            "type": "string",
            "example": "New Chat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "ISO 8601 datetime string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "ISO 8601 datetime string"
          },
          "config": {
            "$ref": "#/components/schemas/ChatConfig"
          },
          "messageCount": {
            "type": "integer",
            "minimum": 0
          },
          "lastMessagePreview": {
            "type": [
              "string",
              "null"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pinned": {
            "type": "boolean"
          },
          "archived": {
            "type": "boolean"
          },
          "icon": {
            "$ref": "#/components/schemas/ChatIcon"
          },
          "folderId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "example": "01935c4a-1b2d-7e3f-8c9d-0e1f2a3b4c5d",
            "description": "UUIDv7 identifier"
          },
          "projectId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "example": "01935c4a-1b2d-7e3f-8c9d-0e1f2a3b4c5d",
            "description": "UUIDv7 identifier"
          },
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "contextRef": {
            "$ref": "#/components/schemas/ChatContextRef"
          }
        },
        "required": [
          "id",
          "workspaceId",
          "title",
          "createdAt",
          "updatedAt",
          "config",
          "messageCount",
          "lastMessagePreview",
          "tags",
          "pinned",
          "archived",
          "icon",
          "folderId",
          "projectId"
        ]
      },
      "ChatConfig": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "openai",
              "openrouter",
              "ollama",
              "anthropic",
              "groq",
              "local"
            ],
            "example": "openai",
            "description": "AI provider identifier"
          },
          "model": {
            "type": "string",
            "example": "gpt-4o"
          },
          "systemPrompt": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "provider",
          "model",
          "systemPrompt"
        ]
      },
      "ChatIcon": {
        "type": [
          "object",
          "null"
        ],
        "properties": {
          "name": {
            "type": "string",
            "enum": [
              "message-square",
              "bot",
              "terminal",
              "book-open",
              "sparkles",
              "lightbulb"
            ]
          },
          "color": {
            "type": "string",
            "enum": [
              "sky",
              "violet",
              "amber",
              "emerald",
              "rose",
              "slate"
            ]
          }
        },
        "required": [
          "name",
          "color"
        ]
      },
      "ChatContextRef": {
        "type": [
          "object",
          "null"
        ],
        "properties": {
          "type": {
            "type": "string",
            "example": "knowledge_page"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "01935c4a-1b2d-7e3f-8c9d-0e1f2a3b4c5d",
            "description": "UUIDv7 identifier"
          }
        },
        "required": [
          "type",
          "id"
        ]
      },
      "ChatResponse": {
        "type": "object",
        "properties": {
          "chat": {
            "$ref": "#/components/schemas/ChatMetadata"
          }
        },
        "required": [
          "chat"
        ]
      },
      "CreateChatInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "New Chat"
          },
          "config": {
            "type": "object",
            "properties": {
              "provider": {
                "type": "string",
                "enum": [
                  "openai",
                  "openrouter",
                  "ollama",
                  "anthropic",
                  "groq",
                  "local"
                ],
                "example": "openai",
                "description": "AI provider identifier"
              },
              "model": {
                "type": "string",
                "example": "gpt-4o"
              },
              "systemPrompt": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          },
          "icon": {
            "$ref": "#/components/schemas/ChatIcon"
          },
          "category": {
            "type": "string"
          },
          "contextRef": {
            "$ref": "#/components/schemas/ChatContextRef"
          }
        }
      },
      "UpdateChatInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "config": {
            "type": "object",
            "properties": {
              "provider": {
                "type": "string",
                "enum": [
                  "openai",
                  "openrouter",
                  "ollama",
                  "anthropic",
                  "groq",
                  "local"
                ],
                "example": "openai",
                "description": "AI provider identifier"
              },
              "model": {
                "type": "string",
                "example": "gpt-4o"
              },
              "systemPrompt": {
                "type": [
                  "string",
                  "null"
                ]
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pinned": {
            "type": "boolean"
          },
          "archived": {
            "type": "boolean"
          },
          "icon": {
            "$ref": "#/components/schemas/ChatIcon"
          },
          "folderId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "example": "01935c4a-1b2d-7e3f-8c9d-0e1f2a3b4c5d",
            "description": "UUIDv7 identifier"
          },
          "projectId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "example": "01935c4a-1b2d-7e3f-8c9d-0e1f2a3b4c5d",
            "description": "UUIDv7 identifier"
          },
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "contextRef": {
            "$ref": "#/components/schemas/ChatContextRef"
          }
        }
      },
...
      "ProviderInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "enum": [
              "openai",
              "openrouter",
              "ollama",
              "anthropic",
              "groq",
              "local"
            ],
            "example": "openai",
            "description": "AI provider identifier"
          },
          "name": {
            "type": "string",
            "example": "OpenAI"
          },
          "requiresApiKey": {
            "type": "boolean"
          },
          "baseURL": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "id",
          "name",
          "requiresApiKey"
        ]
      },
      "ProviderResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "$ref": "#/components/schemas/ProviderInfo"
          },
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelInfo"
            }
          }
        },
        "required": [
          "provider",
          "models"
        ]
      },
      "ModelInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "gpt-4o"
          },
          "providerId": {
            "type": "string",
            "enum": [
              "openai",
              "openrouter",
              "ollama",
              "anthropic",
              "groq",
              "local"
            ],
            "example": "openai",
            "description": "AI provider identifier"
          },
          "name": {
            "type": "string",
            "example": "GPT-4o"
          },
          "description": {
            "type": "string",
            "example": "Most capable OpenAI model"
          },
          "contextWindow": {
            "type": "integer",
            "minimum": 0
          },
          "capabilities": {
            "$ref": "#/components/schemas/ModelCapabilities"
          }
        },
        "required": [
          "id",
          "providerId",
          "name",
          "description",
          "contextWindow",
          "capabilities"
        ]
      },
      "ModelCapabilities": {
        "type": "object",
        "properties": {
          "supportsVision": {
            "type": "boolean"
          },
          "supportsStreaming": {
            "type": "boolean"
          },
          "supportsReasoning": {
            "type": "boolean"
          }
        },
        "required": [
          "supportsVision",
          "supportsStreaming",
          "supportsReasoning"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/chats": {
      "get": {
        "operationId": "listChats",
        "tags": [
          "Chats"
        ],
        "summary": "List all chats",
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ChatListResponse"
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createChat",
        "tags": [
          "Chats"
        ],
        "summary": "Create a new chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response 201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ChatResponse"
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/chats/{id}": {
      "get": {
        "operationId": "getChat",
        "tags": [
          "Chats"
        ],
        "summary": "Get chat by ID",
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ChatResponse"
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "01935c4a-1b2d-7e3f-8c9d-0e1f2a3b4c5d"
          }
        ]
      },
      "patch": {
        "operationId": "updateChat",
        "tags": [
          "Chats"
        ],
        "summary": "Update chat by ID",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChatInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ChatResponse"
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "01935c4a-1b2d-7e3f-8c9d-0e1f2a3b4c5d"
          }
        ]
      },
      "delete": {
        "operationId": "deleteChat",
        "tags": [
          "Chats"
        ],
        "summary": "Delete chat by ID",
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedResponse"
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "01935c4a-1b2d-7e3f-8c9d-0e1f2a3b4c5d"
          }
        ]
      }
    },
...
    "/api/providers/{id}": {
      "get": {
        "operationId": "getProvider",
        "tags": [
          "Providers"
        ],
        "summary": "Get provider with models",
        "responses": {
          "200": {
            "description": "Response 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProviderResponse"
                    }
                  },
                  "required": [
                    "ok",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "01935c4a-1b2d-7e3f-8c9d-0e1f2a3b4c5d"
          }
        ]
      }
    },
...
  },
...
}